spine config error

+++ upgrade-spine.sh
@@ -27,9 +27,6 @@ while :; do
         dev|-dev|--dev)
                 branch="dev"
         ;;
-	php|-php|--php)
-	branch="php"
-        ;;
         *) break
     esac
     shift
@@ -187,10 +184,13 @@ function pick-version() {
 }
 
 function copyConfig() {
-	sudo cp /usr/local/spine/etc/spine.conf.dist /usr/local/spine/etc/spine.conf
-	sudo sed -i 's/cactiuser/cacti/g' /usr/local/spine/etc/spine.conf
-	sudo systemctl start snmpd
-	sudo systemctl enable snmpd
+	if [ -f /usr/local/spine/etc/spine.conf ]; then
+		printerror confExists
+	else
+		printerror confNotExist
+		sudo cp /usr/local/spine/etc/spine.conf.dist /usr/local/spine/etc/spine.conf
+		sudo sed -i 's/cactiuser/cacti/g' /usr/local/spine/etc/spine.conf
+	fi
 }
 
 case "$1" in
